+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [NEXT] - 2022-06-03
+
+## Changed
+
+* Deprecated the 'status' command, in favour of the new 'stats' command.
+
 ## [6.0.0.rc2] - 2022-04-03
 
 ## Changed